Weight distribution of major tuna catches

Weight distribution of major tuna catches (in percentage of the total number of fishes).

bio_weight_tuna(dataframe, report_year, graph_type = "plot", title = FALSE)

Arguments

dataframe

data.frame expected. 'Csv' or 'output' of the function data_extraction, which must be done before using the bio_weight_tuna() function.

report_year

integer expected. Year of the statistical report.

graph_type

character expected. 'plot' or 'ggplot.' Plot by default.

title

TRUE or FALSE expected. Title for plotly graph_type. False by default.

Value

The function return ggplot R plot.

Details

The input dataframe must contain all these columns for the function to work [see referentials]:

   activity_date | school_type | species_code | size_class | estimated_individual
   -------------------------------------------------------------------------------
   2018          | FOB         | 6            | 45         | 1.97
   2018          | FSC         | 6            | 44         | 1.25
   2018          | FOB         | 6            | 43         | 28.7

Add these columns for an automatic title (optional):

country_code

ocean_code