Annual number of fishing sets

Fishing operations. Annual number of fishing sets on FOB-associated and free-swimming tuna schools.

fishing_activity(
  dataframe,
  graph_type = "plot",
  with_catch = "with",
  title = FALSE
)

Arguments

dataframe

data.frame expected. 'Csv' or 'output' of the function data_extraction, which must be done before using the fishing_activity() function.

graph_type

character expected. 'plot', 'plotly' or 'table'. Plot by default.

with_catch

character expected. 'with' or 'without' depending on whether we want the number of sets with or without catches. With by default.

title

TRUE or FALSE expected. Title for plotly graph_type. False by default.

Value

The function return ggplot R plot.

Details

The input dataframe must contain all these columns for the function to work [see referentials]:

   activity_date | school_type | positive_set | total_set
   ------------------------------------------------------
   2010-03-06    | FOB         | 0            | 0
   2010-12-04    | FOB         | 0            | 0
   2010-05-19    | FOB         | 0            | 0

Add these columns for an automatic title (optional):

country_code

ocean_code

vessel_type_code