Sampled biological variables

Give the number of each biological variable sampled for a given year.

species_biological_variable(
  dataframe,
  data_type,
  graph_type = "table",
  reported_year = NULL,
  start_date = NULL,
  end_date = NULL,
  selected_variable = NULL,
  selected_species = NULL
)

Arguments

dataframe

data.frame expected. Csv or output of the function data_extraction, which must be done before using the species_biological_variable() function.

data_type

character expected. 'tunabio' or 'observe'.

graph_type

character expected. 'ggplot' or 'table'. table by default.

reported_year

integer expected. Write the wanted year of the report

start_date

character expected. if reported_year is not given. Write the start date of the time range of the report.

end_date

character expected. if reported_year is not given. Write the end date of the time range of the report

selected_variable

character expected. 'weight', 'length' or 'sex'. If NULL give all the variable for the given year.

selected_species

character expected. Name the species you want to select. Can be more than one. If NULL give all the species for the given year.

Value

The function return ggplot or table R plot.

Details

The input dataframe frome sql must contain all these columns for the function to work [see referentials]:

   fish_sampling_date | weight  | length | sex | species_code_fao | count
   -------------------------------------------------------------------------------
   2022               | 35.2    | 200    | 4   | SAI              | 1
   2022               |  9.93   |  44    | 2   | LKV              | 1
   2022               |  0.058  |  24    | 4   | REO              | 1